+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Added removal of repeated points in linestrings
+
 ## [0.9.4] - 2024-09-11
 
 ### Changed